backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
google-cloud-dialogflow==2.29.0
googlemaps==4.10.0

